+++ Tests/AspNetControllers.SecuredByDefault/AspNetControllers.SecuredByDefault.Api/Program.cs
@@ -1,0 +1,40 @@
+using AspNetControllers.SecuredByDefault.Api.Configuration;
+using AspNetControllers.SecuredByDefault.Api.Filters;
+using AspNetControllers.SecuredByDefault.Application;
+using AspNetControllers.SecuredByDefault.Infrastructure;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.AspNetCore.Program", Version = "1.0")]
+
+namespace AspNetControllers.SecuredByDefault.Api
+{
+    public class Program
+    {
+        public static void Main(string[] args)
+        {
+            var builder = WebApplication.CreateBuilder(args);
+
+            // Add services to the container.
+            builder.Services.AddControllers(
+                opt =>
+                {
+                    opt.Filters.Add<ExceptionFilter>();
+                });
+            builder.Services.AddApplication(builder.Configuration);
+            builder.Services.ConfigureProblemDetails();
+            builder.Services.AddInfrastructure(builder.Configuration);
+
+            var app = builder.Build();
+
+            // Configure the HTTP request pipeline.
+            app.UseExceptionHandler();
+            app.UseHttpsRedirection();
+            app.UseRouting();
+            app.UseAuthorization();
+            app.MapControllers();
+
+            app.Run();
+        }
+    }
+}