+++ Tests/AspNetControllers.SecuredByDefault/AspNetControllers.SecuredByDefault.Application/Interfaces/ITestService.cs
@@ -1,0 +1,12 @@
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.Contracts.ServiceContract", Version = "1.0")]
+
+namespace AspNetControllers.SecuredByDefault.Application.Interfaces
+{
+    public interface ITestService
+    {
+        Task Operation(CancellationToken cancellationToken = default);
+    }
+}